+++ src/main/java/com/example/demo/dto/chat/ChatMemoryVector.java
@@ -11,7 +11,7 @@ public class ChatMemoryVector {
     private Long agentId;
     private Integer orderIndex;
     private String content;
-    private Integer tokenUsage;
+    private Long tokenUsage;
     private OffsetDateTime createdAt;
     private String role;
     private List<Double> embedding;
